# backend/app/CarDamageReports/Controllers/Schemas/CarDamageReportsInput.php
<?php

namespace App\CarDamageReports\Controllers\Schemas;
/**
 * @OA\Schema(
 *     schema="CarDamageReportsInput",
 *     type="object",
 *     title="Car Damage Report Input",
 *     required={"reporter_name","car_model","license_plate","description","damage_level"},
 *     @OA\Property(property="reporter_name", type="string", example="John Doe"),
 *     @OA\Property(property="car_model", type="string", example="Toyota Corolla"),
 *     @OA\Property(property="license_plate", type="string", example="ABC-1234"),
 *     @OA\Property(property="description", type="string", example="Front bumper dented"),
 *     @OA\Property(property="damage_level", type="string", enum={"minor","moderate","severe"}, example="moderate"),
 *     @OA\Property(property="is_resolved", type="boolean", example=false),
 * )
 */
class CarDamageReportsInput
{
    // Empty class, only for Swagger
}